# Improve event details page performance