--- a/includes/attendee/attendee-repository.php
+++ b/includes/attendee/attendee-repository.php
@@ -108,11 +108,40 @@ class Attendee_Repository {
 	}
 
 	/**
+	 * Retrieve all the attendees of an event.
+	 *
 	 * @return Attendee[] Attendees of the event.
+	 * @throws Exception
 	 */
 	public function get_attendees( int $event_id ): array { // phpcs:ignore Generic.CodeAnalysis.UnusedFunctionParameter.Found
-		// TODO.
-		return array();
+		global $wpdb, $gp_table_prefix;
+
+		// phpcs:disable WordPress.DB.PreparedSQL.InterpolatedNotPrepared
+		// phpcs:disable WordPress.DB.DirectDatabaseQuery.DirectQuery
+		// phpcs:disable WordPress.DB.DirectDatabaseQuery.NoCaching
+		$rows = $wpdb->get_results(
+			$wpdb->prepare(
+				"
+				select event_id, user_id, is_host
+				from {$gp_table_prefix}event_attendees
+				where event_id = %d
+			",
+				array(
+					$event_id,
+				)
+			)
+		);
+		// phpcs:enable
+
+		$attendees = array();
+		foreach ( $rows as $row ) {
+			$attendee = new Attendee( $row->event_id, $row->user_id );
+			if ( '1' === $row->is_host ) {
+				$attendee->mark_as_host();
+			}
+			$attendees[] = $attendee;
+		}
+		return $attendees;
 	}
 
 	/**