--- a/includes/attendee/attendee.php
+++ b/includes/attendee/attendee.php
@@ -8,11 +8,12 @@ class Attendee {
 	private int $event_id;
 	private int $user_id;
 	private bool $is_host;
+	private bool $is_contributor;
 
 	/**
 	 * @throws Exception
 	 */
-	public function __construct( int $event_id, int $user_id ) {
+	public function __construct( int $event_id, int $user_id, bool $is_host = false, bool $is_contributor = false ) {
 		if ( $event_id < 1 ) {
 			throw new Exception( 'invalid event id' );
 		}
@@ -20,9 +21,10 @@ class Attendee {
 			throw new Exception( 'invalid user id' );
 		}
 
-		$this->event_id = $event_id;
-		$this->user_id  = $user_id;
-		$this->is_host  = false;
+		$this->event_id       = $event_id;
+		$this->user_id        = $user_id;
+		$this->is_host        = $is_host;
+		$this->is_contributor = $is_contributor;
 	}
 
 	public function event_id(): int {
@@ -35,6 +37,10 @@ class Attendee {
 
 	public function is_host(): bool {
 		return $this->is_host;
+	}
+
+	public function is_contributor(): bool {
+		return $this->is_contributor;
 	}
 
 	public function mark_as_host(): void {