--- a/includes/attendee/attendee.php
+++ b/includes/attendee/attendee.php
@@ -8,12 +8,16 @@ class Attendee {
 	private int $event_id;
 	private int $user_id;
 	private bool $is_host;
-	private bool $is_contributor;
+
+	/**
+	 * @var string[]
+	 */
+	private array $contributed_locales;
 
 	/**
 	 * @throws Exception
 	 */
-	public function __construct( int $event_id, int $user_id, bool $is_host = false, bool $is_contributor = false ) {
+	public function __construct( int $event_id, int $user_id, bool $is_host = false, array $contributed_locales = array() ) {
 		if ( $event_id < 1 ) {
 			throw new Exception( 'invalid event id' );
 		}
@@ -21,10 +25,10 @@ class Attendee {
 			throw new Exception( 'invalid user id' );
 		}
 
-		$this->event_id       = $event_id;
-		$this->user_id        = $user_id;
-		$this->is_host        = $is_host;
-		$this->is_contributor = $is_contributor;
+		$this->event_id            = $event_id;
+		$this->user_id             = $user_id;
+		$this->is_host             = $is_host;
+		$this->contributed_locales = $contributed_locales;
 	}
 
 	public function event_id(): int {
@@ -40,7 +44,7 @@ class Attendee {
 	}
 
 	public function is_contributor(): bool {
-		return $this->is_contributor;
+		return ! empty( $this->contributed_locales );
 	}
 
 	public function mark_as_host(): void {
@@ -49,5 +53,12 @@ class Attendee {
 
 	public function mark_as_non_host(): void {
 		$this->is_host = false;
+	}
+
+	/**
+	 * @return string[]
+	 */
+	public function contributed_locales(): array {
+		return $this->contributed_locales;
 	}
 }