--- a/includes/routes/event/details.php
+++ b/includes/routes/event/details.php
@@ -45,18 +45,19 @@ class Details_Route extends Route {
 		$event_start       = $event->start();
 		$event_end         = $event->end();
 
-		$attendee          = $this->attendee_repository->get_attendee( $event->id(), $user->ID );
-		$user_is_attending = $attendee instanceof Attendee;
-
 		$stats_calculator   = new Stats_Calculator();
 		$project_repository = new Project_Repository();
+
+		$attendee          = $this->attendee_repository->get_attendee( $event->id(), $user->ID );
+		$user_is_attending = $attendee instanceof Attendee;
+		$contributors      = $stats_calculator->get_contributors( $event->id() );
+		$attendee_repo     = $this->attendee_repository;
+		$hosts             = $this->attendee_repository->get_hosts( $event->id() );
+		$attendees         = $this->attendee_repository->get_attendees_not_contributing( $event->id() );
+		$projects          = $project_repository->get_for_event( $event->id() );
+
 		try {
-			$event_stats   = $stats_calculator->for_event( $event->id() );
-			$contributors  = $stats_calculator->get_contributors( $event->id() );
-			$attendee_repo = $this->attendee_repository;
-			$hosts         = $this->attendee_repository->get_hosts( $event->id() );
-			$attendees     = $this->attendee_repository->get_attendees_not_contributing( $event->id() );
-			$projects      = $project_repository->get_for_event( $event->id() );
+			$event_stats = $stats_calculator->for_event( $event->id() );
 		} catch ( Exception $e ) {
 			// phpcs:ignore WordPress.PHP.DevelopmentFunctions.error_log_error_log
 			error_log( $e );