--- a/includes/routes/event/details.php
+++ b/includes/routes/event/details.php
@@ -50,17 +50,22 @@ class Details_Route extends Route {
 
 		$current_user_attendee      = $this->attendee_repository->get_attendee( $event->id(), $user->ID );
 		$user_is_attending          = $current_user_attendee instanceof Attendee;
-		$contributors               = $this->attendee_repository->get_contributors( $event->id() );
 		$attendee_repo              = $this->attendee_repository;
 		$attendees_not_contributing = $this->attendee_repository->get_attendees_not_contributing( $event->id() );
 		$projects                   = $project_repository->get_for_event( $event->id() );
-
-		$attendees = $this->attendee_repository->get_attendees( $event->id() );
+		$attendees                  = $this->attendee_repository->get_attendees( $event->id() );
 
 		$hosts = array();
 		foreach ( $attendees as $attendee ) {
 			if ( $attendee->is_host() ) {
 				$hosts[] = $attendee;
+			}
+		}
+
+		$contributors = array();
+		foreach ( $attendees as $attendee ) {
+			if ( $attendee->is_contributor() ) {
+				$contributors[] = $attendee;
 			}
 		}