… includes/routes/user/attend-event.php
… includes/stats/stats-calculator.php
… includes/translation/translation-repository.php
… includes/upgrade.php
… templates/event.php
… templates/events-header.php
… tests/attendee/attendee-repository.php
… tests/event/event-capabilities.php
… tests/lib/translation-factory.php
… tests/stats/stats-importer.php
… tests/translation/translation-repository.php
… wporg-gp-translation-events.php