Fixes #225 #208 

This PR reduces the time it takes to load the event details page. It does so by changing the logic so that all data is retrieved from the database, and then passed to the template, so that no queries are being done by the template. This reduces cases of needing to do n+1 queries, and makes it so that there's a single query that retrieves attendees/hosts/contributors.

Additionally, the `Attendee` model was improved so that an instance of `Attendee` now knows whether that attendee is a contributor or not (`$attendee->is_contributor()`). This simplifies things quite a bit IMHO.

Concerning finding who is a new contributor, that is now done in a single query for all contributors of an event, instead of doing one query per contributor. This was the biggest culprit of the performance issues, and where the most gains come from.

Note that the `Attendee` does **not** currently have an `$attendee->is_new_contributor()`. This is because retrieving this information in `get_attendees()` would make that method slow, and we don't necessarily need to know who is a new contributor in every place we call `get_attendees()`. When we do need that information, we can do something like [this](https://github.com/WordPress/wporg-gp-translation-events/pull/236/files#diff-191ce4bb541891d5409a539d712644c17ed150c3b7475e94444fb3a6c00eb2a3R80-R93).

## Metrics

Tested on my wporg sandbox, for the [Yoast Contributor Day](https://translate.wordpress.org/events/yoast-contributor-day-2024-april-edition/) event.

| Metric                       | Before | After |
|------------------------------|--------|-------|
| Page generation time         | 13.3s  | **4.3s**  |
| Query time                   | 10.4s  | **1.6s**  |
| Number of slow queries       | 10     | **3**     |
| Number of duplicated queries | 398    | **314**   |