@psrpinto 2024-05-03
> I think we should consider dropping the count_translations_before function and move the is_new_contributor value to the event_attendees table.

This makes a lot of sense! Opened #241.

I've addressed all your remarks (great ones BTW!), could you take a quick look?